fix : 엔드포인트 url 및 dto 필드 타입 변경

+++ src/main/java/live/ioteatime/apiservice/domain/Organization.java
@@ -3,7 +3,6 @@ package live.ioteatime.apiservice.domain;
 import lombok.Getter;
 
 import javax.persistence.*;
-import java.math.BigInteger;
 
 @Entity
 @Getter
@@ -15,7 +14,7 @@ public class Organization {
     private int id;
     private String name;
     @Column(name = "electricity_budget")
-    private BigInteger electricityBudget;
+    private Long electricityBudget;
     @Column(name = "organization_code")
     private String organizationCode;
 }
